+++ src/main/java/com/phantomwing/thesilverage/datagen/ModItemTagsProvider.java
@@ -8,6 +8,7 @@ import net.minecraft.core.HolderLookup;
 import net.minecraft.data.PackOutput;
 import net.minecraft.data.tags.ItemTagsProvider;
 import net.minecraft.tags.ItemTags;
+import net.minecraft.world.item.Items;
 import net.minecraft.world.level.block.Block;
 import net.neoforged.neoforge.common.Tags;
 import net.neoforged.neoforge.common.data.ExistingFileHelper;
@@ -23,26 +24,19 @@ public class ModItemTagsProvider extends ItemTagsProvider {
 
     @Override
     protected void addTags(HolderLookup.@NotNull Provider provider) {
-        // Tools
-        tag(ItemTags.SWORDS).add(ModItems.SILVER_SWORD.get());
-        tag(ItemTags.SHOVELS).add(ModItems.SILVER_SHOVEL.get());
-        tag(ItemTags.PICKAXES).add(ModItems.SILVER_PICKAXE.get());
-        tag(ItemTags.AXES).add(ModItems.SILVER_AXE.get());
-        tag(ItemTags.HOES).add(ModItems.SILVER_HOE.get());
-        tag(ModTags.Items.SILVER_TOOL_MATERIALS).add(ModItems.SILVER_INGOT.get());
+        addModTags(provider);
+        addCommonTags(provider);
+        addMinecraftTags(provider);
+    }
 
-        // Armor
-        tag(ItemTags.HEAD_ARMOR).add(ModItems.SILVER_HELMET.get());
-        tag(ItemTags.CHEST_ARMOR).add(ModItems.SILVER_CHESTPLATE.get());
-        tag(ItemTags.LEG_ARMOR).add(ModItems.SILVER_LEGGINGS.get());
-        tag(ItemTags.FOOT_ARMOR).add(ModItems.SILVER_BOOTS.get());
-        tag(ItemTags.TRIMMABLE_ARMOR).add(
-                ModItems.SILVER_HELMET.get(),
-                ModItems.SILVER_CHESTPLATE.get(),
-                ModItems.SILVER_LEGGINGS.get(),
-                ModItems.SILVER_BOOTS.get()
-        );
+    protected void addModTags(HolderLookup.@NotNull Provider provider) {
+        // Glistering nuggets (used for e.g. the Glistering Melon override recipe)
+        tag(ModTags.Items.GLISTERING_NUGGETS)
+                .add(Items.GOLD_NUGGET)
+                .add(ModItems.SILVER_NUGGET.get());
+    }
 
+    protected void addCommonTags(HolderLookup.@NotNull Provider provider) {
         // Storage blocks
         tag(CommonTags.Items.STORAGE_BLOCKS_SILVER).add(ModItems.SILVER_BLOCK.get());
         tag(CommonTags.Items.STORAGE_BLOCKS_RAW_SILVER).add(ModItems.RAW_SILVER_BLOCK.get());
@@ -53,6 +47,9 @@ public class ModItemTagsProvider extends ItemTagsProvider {
         // Raw materials
         tag(CommonTags.Items.RAW_MATERIALS_SILVER).add(ModItems.RAW_SILVER.get());
         tag(Tags.Items.RAW_MATERIALS).addTag(CommonTags.Items.RAW_MATERIALS_SILVER);
+
+        // Materials
+        tag(CommonTags.Items.TOOL_MATERIALS_SILVER).add(ModItems.SILVER_INGOT.get());
 
         // Nuggets
         tag(CommonTags.Items.NUGGETS_SILVER).add(ModItems.SILVER_NUGGET.get());
@@ -67,10 +64,33 @@ public class ModItemTagsProvider extends ItemTagsProvider {
         tag(Tags.Items.ORES).addTag(CommonTags.Items.ORES_SILVER);
         tag(Tags.Items.ORES_IN_GROUND_STONE).add(ModItems.SILVER_ORE.get());
         tag(Tags.Items.ORES_IN_GROUND_DEEPSLATE).add(ModItems.DEEPSLATE_SILVER_ORE.get());
+    }
 
-        // Silver misc
-        tag(ItemTags.BEACON_PAYMENT_ITEMS).add(ModItems.SILVER_INGOT.get());
+    protected void addMinecraftTags(HolderLookup.@NotNull Provider provider) {
+        // Tools
+        tag(ItemTags.SWORDS).add(ModItems.SILVER_SWORD.get());
+        tag(ItemTags.SHOVELS).add(ModItems.SILVER_SHOVEL.get());
+        tag(ItemTags.PICKAXES).add(ModItems.SILVER_PICKAXE.get());
+        tag(ItemTags.AXES).add(ModItems.SILVER_AXE.get());
+        tag(ItemTags.HOES).add(ModItems.SILVER_HOE.get());
+
+        // Armor
+        tag(ItemTags.HEAD_ARMOR).add(ModItems.SILVER_HELMET.get());
+        tag(ItemTags.CHEST_ARMOR).add(ModItems.SILVER_CHESTPLATE.get());
+        tag(ItemTags.LEG_ARMOR).add(ModItems.SILVER_LEGGINGS.get());
+        tag(ItemTags.FOOT_ARMOR).add(ModItems.SILVER_BOOTS.get());
+        tag(ItemTags.TRIMMABLE_ARMOR).add(
+                ModItems.SILVER_HELMET.get(),
+                ModItems.SILVER_CHESTPLATE.get(),
+                ModItems.SILVER_LEGGINGS.get(),
+                ModItems.SILVER_BOOTS.get()
+        );
+
+        // Armor trims
         tag(ItemTags.TRIM_MATERIALS).add(ModItems.SILVER_INGOT.get());
+
+        // Beacon
+        tag(ItemTags.BEACON_PAYMENT_ITEMS).add(ModItems.SILVER_INGOT.get());
 
         // Slabs
         tag(ItemTags.SLABS)
@@ -83,6 +103,7 @@ public class ModItemTagsProvider extends ItemTagsProvider {
                 .add(ModItems.WAXED_WEATHERED_CUT_SILVER_SLAB.get())
                 .add(ModItems.WAXED_OXIDIZED_CUT_SILVER_SLAB.get());
 
+        // Stairs
         tag(ItemTags.STAIRS)
                 .add(ModItems.CUT_SILVER_STAIRS.get())
                 .add(ModItems.EXPOSED_CUT_SILVER_STAIRS.get())
@@ -93,6 +114,7 @@ public class ModItemTagsProvider extends ItemTagsProvider {
                 .add(ModItems.WAXED_WEATHERED_CUT_SILVER_STAIRS.get())
                 .add(ModItems.WAXED_OXIDIZED_CUT_SILVER_STAIRS.get());
 
+        // Trapdoors
         tag(ItemTags.TRAPDOORS)
                 .add(ModItems.SILVER_TRAPDOOR.get())
                 .add(ModItems.EXPOSED_SILVER_TRAPDOOR.get())
@@ -103,6 +125,7 @@ public class ModItemTagsProvider extends ItemTagsProvider {
                 .add(ModItems.WAXED_WEATHERED_SILVER_TRAPDOOR.get())
                 .add(ModItems.WAXED_OXIDIZED_SILVER_TRAPDOOR.get());
 
+        // Doors
         tag(ItemTags.DOORS)
                 .add(ModItems.SILVER_DOOR.get())
                 .add(ModItems.EXPOSED_SILVER_DOOR.get())