+++ src/main/java/com/phantomwing/thesilverage/datagen/ModRecipeProvider.java
@@ -1,7 +1,10 @@
 package com.phantomwing.thesilverage.datagen;
 
+import com.phantomwing.thesilverage.Configuration;
 import com.phantomwing.thesilverage.TheSilverAge;
+import com.phantomwing.thesilverage.condition.ConfigBooleanCondition;
 import com.phantomwing.thesilverage.item.ModItems;
+import com.phantomwing.thesilverage.tags.ModTags;
 import com.phantomwing.thesilverage.utils.ItemUtils;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.data.PackOutput;
@@ -9,10 +12,15 @@ import net.minecraft.data.recipes.*;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.item.crafting.*;
 import net.minecraft.world.level.ItemLike;
+import net.neoforged.neoforge.common.conditions.ICondition;
+import net.neoforged.neoforge.common.conditions.NotCondition;
 import org.jetbrains.annotations.NotNull;
 import java.util.concurrent.CompletableFuture;
 
 public class ModRecipeProvider extends RecipeProvider {
+    private static final float XP_TINY = 0.1f;
+    private static final float XP_MEDIUM = 1f;
+
     public ModRecipeProvider(PackOutput output, CompletableFuture<HolderLookup.Provider> lookupProvider) {
         super(output, lookupProvider);
     }
@@ -20,22 +28,41 @@ public class ModRecipeProvider extends RecipeProvider {
     @Override
     protected void buildRecipes(@NotNull RecipeOutput output) {
         buildCraftingRecipes(output);
+        buildRecipeOverrides(output);
     }
 
     private void buildCraftingRecipes(@NotNull RecipeOutput output) {
-        oreSmeltingRecipes(output, ModItems.RAW_SILVER, ModItems.SILVER_INGOT, 1f);
+        oreSmeltingRecipes(output, ModItems.RAW_SILVER, ModItems.SILVER_INGOT, XP_MEDIUM);
 
         // Storage item recipes
         storageItemRecipes(output, RecipeCategory.MISC, ModItems.SILVER_NUGGET, ModItems.SILVER_INGOT);
         storageItemRecipes(output, RecipeCategory.MISC, ModItems.RAW_SILVER, ModItems.RAW_SILVER_BLOCK);
 
         // Tools
-        sword(output, ModItems.SILVER_INGOT, ModItems.SILVER_SWORD);
-        pickaxe(output, ModItems.SILVER_INGOT, ModItems.SILVER_PICKAXE);
-        axe(output, ModItems.SILVER_INGOT, ModItems.SILVER_AXE);
-        hoe(output, ModItems.SILVER_INGOT, ModItems.SILVER_HOE);
-        shovel(output, ModItems.SILVER_INGOT, ModItems.SILVER_SHOVEL);
+        sword(output, ModItems.SILVER_SWORD, ModItems.SILVER_INGOT);
+        pickaxe(output, ModItems.SILVER_PICKAXE, ModItems.SILVER_INGOT);
+        axe(output, ModItems.SILVER_AXE, ModItems.SILVER_INGOT);
+        hoe(output, ModItems.SILVER_HOE, ModItems.SILVER_INGOT);
+        shovel(output, ModItems.SILVER_SHOVEL, ModItems.SILVER_INGOT);
 
+        // Armor
+        helmet(output, ModItems.SILVER_HELMET, ModItems.SILVER_INGOT);
+        chestplate(output, ModItems.SILVER_CHESTPLATE, ModItems.SILVER_INGOT);
+        leggings(output, ModItems.SILVER_LEGGINGS, ModItems.SILVER_INGOT);
+        boots(output, ModItems.SILVER_BOOTS, ModItems.SILVER_INGOT);
+
+        // Smelting tools/armor into nuggets
+        oreSmeltingRecipes(output, ModItems.SILVER_SWORD, ModItems.SILVER_NUGGET, XP_TINY);
+        oreSmeltingRecipes(output, ModItems.SILVER_PICKAXE, ModItems.SILVER_NUGGET, XP_TINY);
+        oreSmeltingRecipes(output, ModItems.SILVER_AXE, ModItems.SILVER_NUGGET, XP_TINY);
+        oreSmeltingRecipes(output, ModItems.SILVER_HOE, ModItems.SILVER_NUGGET, XP_TINY);
+        oreSmeltingRecipes(output, ModItems.SILVER_SHOVEL, ModItems.SILVER_NUGGET, XP_TINY);
+        oreSmeltingRecipes(output, ModItems.SILVER_HELMET, ModItems.SILVER_NUGGET, XP_TINY);
+        oreSmeltingRecipes(output, ModItems.SILVER_CHESTPLATE, ModItems.SILVER_NUGGET, XP_TINY);
+        oreSmeltingRecipes(output, ModItems.SILVER_LEGGINGS, ModItems.SILVER_NUGGET, XP_TINY);
+        oreSmeltingRecipes(output, ModItems.SILVER_BOOTS, ModItems.SILVER_NUGGET, XP_TINY);
+        oreSmeltingRecipes(output, ModItems.SILVER_HORSE_ARMOR, ModItems.SILVER_NUGGET, XP_TINY);
+        
         // Block of Silver
         storageItemRecipes(output, RecipeCategory.MISC, ModItems.SILVER_INGOT, ModItems.SILVER_BLOCK);
         waxable(output, ModItems.SILVER_BLOCK, ModItems.WAXED_SILVER_BLOCK);
@@ -188,6 +215,111 @@ public class ModRecipeProvider extends RecipeProvider {
         waxable(output, ModItems.OXIDIZED_SILVER_TRAPDOOR, ModItems.WAXED_OXIDIZED_SILVER_TRAPDOOR);
     }
 
+    /** Add overrides for Vanilla Minecraft recipes. (Only if a recipe is enabled) */
+    private void buildRecipeOverrides(@NotNull RecipeOutput output) {
+        ICondition condition = new ConfigBooleanCondition(Configuration.OVERRIDE_VANILLA_RECIPES_ID);
+        RecipeOutput conditionalOutput = output.withConditions(condition);
+        RecipeOutput fallbackOutput = output.withConditions(new NotCondition(condition));
+
+        // Lodestone
+        ShapedRecipeBuilder.shaped(RecipeCategory.MISC, Items.LODESTONE, 1)
+                .pattern("###")
+                .pattern("#S#")
+                .pattern("###")
+                .define('#', Items.CHISELED_STONE_BRICKS)
+                .define('S', ModItems.SILVER_INGOT)
+                .unlockedBy(getHasName(ModItems.SILVER_INGOT), has(ModItems.SILVER_INGOT))
+                .save(conditionalOutput);
+        ShapedRecipeBuilder.shaped(RecipeCategory.MISC, Items.LODESTONE, 1)
+                .pattern("###")
+                .pattern("#S#")
+                .pattern("###")
+                .define('#', Items.CHISELED_STONE_BRICKS)
+                .define('S', Items.NETHERITE_INGOT)
+                .unlockedBy(getHasName(Items.NETHERITE_INGOT), has(Items.NETHERITE_INGOT)) // TODO: Change fallback to IRON_INGOT once updated to later Minecraft version
+                .save(fallbackOutput, "minecraft:" + ItemUtils.getName(Items.LODESTONE) + "_fallback"); // Original recipe if override is disabled
+
+        // Brewing Stand
+        ShapedRecipeBuilder.shaped(RecipeCategory.MISC, Items.BREWING_STAND, 1)
+                .pattern(" B ")
+                .pattern("SSS")
+                .define('B', Items.BLAZE_ROD)
+                .define('S', ModItems.SILVER_INGOT)
+                .unlockedBy(getHasName(Items.BLAZE_ROD), has(Items.BLAZE_ROD))
+                .save(conditionalOutput);
+        ShapedRecipeBuilder.shaped(RecipeCategory.MISC, Items.BREWING_STAND, 1)
+                .pattern(" B ")
+                .pattern("SSS")
+                .define('B', Items.BLAZE_ROD)
+                .define('S', Items.STONE)
+                .unlockedBy(getHasName(Items.BLAZE_ROD), has(Items.BLAZE_ROD))
+                .save(fallbackOutput, "minecraft:" + ItemUtils.getName(Items.BREWING_STAND) + "_fallback");  // Original recipe if override is disabled
+
+        // Glistering Melon Slice
+        ShapedRecipeBuilder.shaped(RecipeCategory.MISC, Items.GLISTERING_MELON_SLICE, 1)
+                .pattern("###")
+                .pattern("#M#")
+                .pattern("###")
+                .define('#', ModTags.Items.GLISTERING_NUGGETS)
+                .define('M', Items.MELON_SLICE)
+                .unlockedBy(getHasName(Items.MELON_SLICE), has(Items.MELON_SLICE))
+                .save(conditionalOutput);
+        ShapedRecipeBuilder.shaped(RecipeCategory.MISC, Items.GLISTERING_MELON_SLICE, 1)
+                .pattern("###")
+                .pattern("#M#")
+                .pattern("###")
+                .define('#', Items.GOLD_NUGGET)
+                .define('M', Items.MELON_SLICE)
+                .unlockedBy(getHasName(Items.MELON_SLICE), has(Items.MELON_SLICE))
+                .save(fallbackOutput, "minecraft:" + ItemUtils.getName(Items.GLISTERING_MELON_SLICE) + "_fallback");  // Original recipe if override is disabled
+
+        // Name Tag
+        ShapelessRecipeBuilder.shapeless(RecipeCategory.MISC, Items.NAME_TAG, 1)
+                .requires(Items.PAPER)
+                .requires(ModItems.SILVER_NUGGET)
+                .unlockedBy(getHasName(Items.PAPER), has(Items.PAPER))
+                .save(conditionalOutput);
+        // TODO: Add fallback recipe when Name Tag recipe is added in later Minecraft version
+
+        // Redstone Comparator
+        ShapedRecipeBuilder.shaped(RecipeCategory.REDSTONE, Items.COMPARATOR, 1)
+                .pattern(" T ")
+                .pattern("TQT")
+                .pattern("SSS")
+                .define('T', Items.REDSTONE_TORCH)
+                .define('Q', Items.QUARTZ)
+                .define('S', ModItems.SILVER_INGOT)
+                .unlockedBy(getHasName(Items.REDSTONE_TORCH), has(Items.REDSTONE_TORCH))
+                .save(conditionalOutput);
+        ShapedRecipeBuilder.shaped(RecipeCategory.REDSTONE, Items.COMPARATOR, 1)
+                .pattern(" T ")
+                .pattern("TQT")
+                .pattern("SSS")
+                .define('T', Items.REDSTONE_TORCH)
+                .define('Q', Items.QUARTZ)
+                .define('S', Items.STONE)
+                .unlockedBy(getHasName(Items.REDSTONE_TORCH), has(Items.REDSTONE_TORCH))
+                .save(fallbackOutput, "minecraft:" + ItemUtils.getName(Items.COMPARATOR) + "_fallback");  // Original recipe if override is disabled
+
+        // Redstone Repeater
+        ShapedRecipeBuilder.shaped(RecipeCategory.REDSTONE, Items.REPEATER, 1)
+                .pattern("TRT")
+                .pattern("SSS")
+                .define('R', Items.REDSTONE)
+                .define('T', Items.REDSTONE_TORCH)
+                .define('S', ModItems.SILVER_INGOT)
+                .unlockedBy(getHasName(Items.REDSTONE), has(Items.REDSTONE))
+                .save(conditionalOutput);
+        ShapedRecipeBuilder.shaped(RecipeCategory.REDSTONE, Items.REPEATER, 1)
+                .pattern("TRT")
+                .pattern("SSS")
+                .define('R', Items.REDSTONE)
+                .define('T', Items.REDSTONE_TORCH)
+                .define('S', Items.STONE)
+                .unlockedBy(getHasName(Items.REDSTONE), has(Items.REDSTONE))
+                .save(fallbackOutput, "minecraft:" + ItemUtils.getName(Items.REPEATER) + "_fallback");  // Original recipe if override is disabled
+    }
+
     private static void stairsWithCutting(RecipeOutput recipeOutput, ItemLike item, ItemLike material) {
         stoneCutting(recipeOutput, item, material, 1);
         stairs(recipeOutput, item, material);
@@ -296,7 +428,7 @@ public class ModRecipeProvider extends RecipeProvider {
                 .save(recipeOutput, getRecipeName(item, result));
     }
 
-    protected static void sword(RecipeOutput recipeOutput, ItemLike material, ItemLike tool) {
+    protected static void sword(RecipeOutput recipeOutput, ItemLike tool, ItemLike material) {
         ShapedRecipeBuilder.shaped(RecipeCategory.COMBAT, tool)
             .pattern("#")
             .pattern("#")
@@ -307,7 +439,7 @@ public class ModRecipeProvider extends RecipeProvider {
             .save(recipeOutput);
     }
 
-    protected static void pickaxe(RecipeOutput recipeOutput, ItemLike material, ItemLike tool) {
+    protected static void pickaxe(RecipeOutput recipeOutput, ItemLike tool, ItemLike material) {
         ShapedRecipeBuilder.shaped(RecipeCategory.COMBAT, tool)
             .pattern("###")
             .pattern(" S ")
@@ -317,7 +449,7 @@ public class ModRecipeProvider extends RecipeProvider {
             .unlockedBy(getHasName(material), has(material))
             .save(recipeOutput);
     }
-    protected static void axe(RecipeOutput recipeOutput, ItemLike material, ItemLike tool) {
+    protected static void axe(RecipeOutput recipeOutput, ItemLike tool, ItemLike material) {
         ShapedRecipeBuilder.shaped(RecipeCategory.COMBAT, tool)
             .pattern("##")
             .pattern("#S")
@@ -328,7 +460,7 @@ public class ModRecipeProvider extends RecipeProvider {
             .save(recipeOutput);
     }
 
-    protected static void hoe(RecipeOutput recipeOutput, ItemLike material, ItemLike tool) {
+    protected static void hoe(RecipeOutput recipeOutput, ItemLike tool, ItemLike material) {
         ShapedRecipeBuilder.shaped(RecipeCategory.COMBAT, tool)
             .pattern("##")
             .pattern(" S")
@@ -339,7 +471,7 @@ public class ModRecipeProvider extends RecipeProvider {
             .save(recipeOutput);
     }
 
-    protected static void shovel(RecipeOutput recipeOutput, ItemLike material, ItemLike tool) {
+    protected static void shovel(RecipeOutput recipeOutput, ItemLike tool, ItemLike material) {
         ShapedRecipeBuilder.shaped(RecipeCategory.COMBAT, tool)
             .pattern("#")
             .pattern("S")
@@ -350,6 +482,43 @@ public class ModRecipeProvider extends RecipeProvider {
             .save(recipeOutput);
     }
 
+    protected static void helmet(RecipeOutput recipeOutput, ItemLike helmet, ItemLike material) {
+        ShapedRecipeBuilder.shaped(RecipeCategory.COMBAT, helmet)
+                .pattern("###")
+                .pattern("# #")
+                .define('#', material)
+                .unlockedBy(getHasName(material), has(material))
+                .save(recipeOutput);
+    }
+
+    protected static void chestplate(RecipeOutput recipeOutput, ItemLike chestplate, ItemLike material) {
+        ShapedRecipeBuilder.shaped(RecipeCategory.COMBAT, chestplate)
+                .pattern("# #")
+                .pattern("###")
+                .pattern("###")
+                .define('#', material)
+                .unlockedBy(getHasName(material), has(material))
+                .save(recipeOutput);
+    }
+
+    protected static void leggings(RecipeOutput recipeOutput, ItemLike leggings, ItemLike material) {
+        ShapedRecipeBuilder.shaped(RecipeCategory.COMBAT, leggings)
+                .pattern("###")
+                .pattern("# #")
+                .pattern("# #")
+                .define('#', material)
+                .unlockedBy(getHasName(material), has(material))
+                .save(recipeOutput);
+    }
+
+    protected static void boots(RecipeOutput recipeOutput, ItemLike boots, ItemLike material) {
+        ShapedRecipeBuilder.shaped(RecipeCategory.COMBAT, boots)
+                .pattern("# #")
+                .pattern("# #")
+                .define('#', material)
+                .unlockedBy(getHasName(material), has(material))
+                .save(recipeOutput);
+    }
 
     protected static void oreSmeltingRecipes(@NotNull RecipeOutput recipeOutput, @NotNull ItemLike material, @NotNull ItemLike result, float experience) {
         smelting(recipeOutput, RecipeCategory.MISC, material, result, experience, 200);
@@ -366,14 +535,14 @@ public class ModRecipeProvider extends RecipeProvider {
         SimpleCookingRecipeBuilder
                 .generic(Ingredient.of(material), category, result, experience, cookingTime, RecipeSerializer.SMELTING_RECIPE, SmeltingRecipe::new)
                 .unlockedBy(getHasName(material), has(material))
-                .save(recipeOutput);
+                .save(recipeOutput, ItemUtils.getNameWithNamespace(result) + "_from_" + ItemUtils.getName(material) + "_smelting");
     }
 
     protected static void blasting(@NotNull RecipeOutput recipeOutput, RecipeCategory category, @NotNull ItemLike material, @NotNull ItemLike result, float experience, int cookingTime) {
         SimpleCookingRecipeBuilder
                 .generic(Ingredient.of(material), category, result, experience, cookingTime, RecipeSerializer.BLASTING_RECIPE, BlastingRecipe::new)
                 .unlockedBy(getHasName(material), has(material))
-                .save(recipeOutput, ItemUtils.getNameWithNamespace(result) + "_from_blasting");
+                .save(recipeOutput, ItemUtils.getNameWithNamespace(result) + "_from_" + ItemUtils.getName(material) + "_blasting");
     }
 
     protected static void smoking(@NotNull RecipeOutput recipeOutput, RecipeCategory category, @NotNull ItemLike material, @NotNull ItemLike result, float experience, int cookingTime) {