+++ changelog.md
@@ -1,4 +1,29 @@
 # 1.0.0
 
-### Additions
-- Initial release of the project.
+Initial release of the project.
+
+Notable features:
+### Silver
+- Silver Ore generation added to Overworld biomes
+  - More common in Deep Dark / Ancient City biomes
+- Oxidization mechanic for Silver blocks
+  - Block of Silver, Cut Silver, Cut Silver Slab, Cut Silver Stairs, Chiseled Silver, Silver Grate, Silver Door, Silver Trapdoor
+- Silver Tools & Armor
+  - Silver Sword, Silver Pickaxe, Silver Axe, Silver Shovel, Silver Hoe
+  - Silver Helmet, Silver Chestplate, Silver Leggings, Silver Boots
+  - Silver Horse Armor
+  - Use Silver Ingot to repair your tools & armor
+  - Use Silver Ingot as an armor trim material
+- Use Silver blocks/ingots to power a Beacon
+- Use Silver Nuggets when crafting a Firework Star
+- Silverfish now have a chance to drop Silver Nuggets when defeated (Configurable!)
+
+### Vanilla Minecraft Changes
+- Rebalanced Minecraft crafting recipes to require Silver (Configurable!)
+  - Brewing Stand
+  - Redstone Repeater
+  - Redstone Comparator
+  - Name Tag
+  - Lodestone
+  - Glistering Melon Slice (now crafted with Silver Nuggets or Gold Nuggets)
+- Some textures updated to include Silver details